retirado system.drawing, convertido para skiasharp
