+++ BoletoNetCore.QuestPdf/BoletoNetCore.QuestPdf.csproj
@@ -7,7 +7,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="5.0.10" />
     <PackageReference Include="QuestPDF" Version="2021.9.3" />
-    <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="2.80.3" />
+    <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="2.88.3" />
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
     <PackageReference Include="Barcoder" Version="1.0.0.90" />
     <PackageReference Include="Barcoder.Renderer.Image" Version="1.0.0.90" />
