+++ BoletoNetCore.Testes/BancoItauCarteira109Tests.cs
@@ -22,7 +22,7 @@ namespace BoletoNetCore.Testes
                 TipoImpressaoBoleto = TipoImpressaoBoleto.Empresa
             };
             _banco = Banco.Instancia(Bancos.Itau);
-            _banco.Beneficiario = Utils.GerarBeneficiario("", "", "", contaBancaria);
+            _banco.Beneficiario = TestUtils.GerarBeneficiario("", "", "", contaBancaria);
 
             _banco.FormataBeneficiario();
         }
@@ -30,13 +30,13 @@ namespace BoletoNetCore.Testes
         [Test]
         public void Itau_109_REM400()
         {
-            Utils.TestarHomologacao(_banco, TipoArquivo.CNAB400, nameof(BancoItauCarteira109Tests), 5, true, "N", 223344);
+            TestUtils.TestarHomologacao(_banco, TipoArquivo.CNAB400, nameof(BancoItauCarteira109Tests), 5, true, "N", 223344);
         }
 
         [Test]
         public void Itau_109_REM240()
         {
-            Utils.TestarHomologacao(_banco, TipoArquivo.CNAB240, nameof(BancoItauCarteira109Tests), 5, true, "N", 223344);
+            TestUtils.TestarHomologacao(_banco, TipoArquivo.CNAB240, nameof(BancoItauCarteira109Tests), 5, true, "N", 223344);
         }
 
         [TestCase(600.00, "223350", "BB000001F", "1", "109/00223350-2", "34191719500000600001090022335021234567890000", "34191.09008 22335.021238 45678.900007 1 71950000060000", 2017, 06, 19)]
@@ -59,7 +59,7 @@ namespace BoletoNetCore.Testes
                 NossoNumero = nossoNumero,
                 NumeroDocumento = numeroDocumento,
                 EspecieDocumento = TipoEspecieDocumento.DM,
-                Pagador = Utils.GerarPagador()
+                Pagador = TestUtils.GerarPagador()
             };
 
             //Ação
