+++ BoletoNetCore/BoletoNetCore.csproj
@@ -26,6 +26,7 @@
 
   <ItemGroup>
     <PackageReference Include="SkiaSharp" Version="2.88.3" />
+    <PackageReference Include="SkiaSharp.NativeAssets.Linux" Version="2.88.3" />
     <PackageReference Include="System.Net.Http.Json" Version="5.0.0" />
   </ItemGroup>
